Puzzle Manager for Diving Bell Level
The logic for the PuzzleManager was completed, allowing it to be completed when all columns were rotated correctly. In addition, a crazy logic was created to adapt an array of Sound Cues to Audio Components. Finally, the logic to activate and deactivate the portal was added, although it does not work yet.