Second level "Dystopia" added
A map was added for the second level called Dystopia. Lighting, post-processing, a landscape, and, for now, a fractal mesh were added. It has no logic for now.
Logic was also added so the Diving Bell portal sends the player to the second level when it collides with it after being activated.