PressurePlates Logic implemented
Logic was added for the second level puzzle based on pressure plates, detecting if there is a collision on it and thus increasing the counter of plates pressed to the BP_Gate, which compares if the plates currently pressed are the ones needed to raise the gate, and if not, returns it to its original position. It was adapted so that the number of plates can be adjusted from the editor.